Filtering: copying VCF instead of symlingking

+++ source/variants/filtering.py
@@ -482,8 +482,9 @@ def postprocess_vcf(sample, anno_vcf_fpath, work_filt_vcf_fpath):
 
     # Moving final VCF
     if isfile(final_vcf_fpath): os.remove(final_vcf_fpath)
-    shutil.move(work_filt_vcf_fpath, final_vcf_fpath)
-    os.symlink(final_vcf_fpath, work_filt_vcf_fpath)
+    # shutil.move(work_filt_vcf_fpath, final_vcf_fpath)
+    # os.symlink(final_vcf_fpath, work_filt_vcf_fpath)
+    shutil.copy(work_filt_vcf_fpath, final_vcf_fpath)
 
     igvtools_index(cnf, final_vcf_fpath)
 
@@ -501,8 +502,7 @@ def postprocess_vcf(sample, anno_vcf_fpath, work_filt_vcf_fpath):
     if work_filt_vcf_fpath and 'tsv_fields' in cnf:
         tsv_fpath = make_tsv(cnf, work_filt_vcf_fpath)
 
-        if isfile(final_tsv_fpath):
-            os.remove(final_tsv_fpath)
+        if isfile(final_tsv_fpath): os.remove(final_tsv_fpath)
         shutil.move(tsv_fpath, final_tsv_fpath)
     else:
         final_tsv_fpath = None
@@ -523,8 +523,7 @@ def postprocess_vcf(sample, anno_vcf_fpath, work_filt_vcf_fpath):
                                    tumor_sample_name=sample.name,
                                    bam_fpath=sample.bam,
                                    normal_sample_name=sample.normal_match.name if sample.normal_match else None)
-        if isfile(final_maf_fpath):
-            os.remove(final_maf_fpath)
+        if isfile(final_maf_fpath): os.remove(final_maf_fpath)
         shutil.move(maf_fpath, final_maf_fpath)
         info('-' * 70)
         info()
